+++ crates/matrix-sdk/src/media.rs
@@ -347,7 +347,7 @@ impl Media {
     /// * `use_cache` - If we should use the media cache for this file.
     pub async fn get_file(
         &self,
-        event_content: impl MediaEventContent,
+        event_content: &impl MediaEventContent,
         use_cache: bool,
     ) -> Result<Option<Vec<u8>>> {
         let Some(source) = event_content.source() else { return Ok(None) };
@@ -365,7 +365,7 @@ impl Media {
     /// # Arguments
     ///
     /// * `event_content` - The media event content.
-    pub async fn remove_file(&self, event_content: impl MediaEventContent) -> Result<()> {
+    pub async fn remove_file(&self, event_content: &impl MediaEventContent) -> Result<()> {
         if let Some(source) = event_content.source() {
             self.remove_media_content(&MediaRequest { source, format: MediaFormat::File }).await?;
         }
@@ -393,7 +393,7 @@ impl Media {
     /// * `use_cache` - If we should use the media cache for this thumbnail.
     pub async fn get_thumbnail(
         &self,
-        event_content: impl MediaEventContent,
+        event_content: &impl MediaEventContent,
         size: MediaThumbnailSize,
         use_cache: bool,
     ) -> Result<Option<Vec<u8>>> {
@@ -420,7 +420,7 @@ impl Media {
     ///   requested with [`get_thumbnail`](#method.get_thumbnail).
     pub async fn remove_thumbnail(
         &self,
-        event_content: impl MediaEventContent,
+        event_content: &impl MediaEventContent,
         size: MediaThumbnailSize,
     ) -> Result<()> {
         if let Some(source) = event_content.source() {